Rename plans for demonstration
openSUSE is also removed from the UI since it still is busted even
with DHCP6 as I have it right now.

For some very uninteresting time-related reasons for demonstrations,
and given the pre-existing terminology choice of "vcpu," I presume all
cores we sell are 2:1 hyperthreaded here, and update the tests
accordingly.

But longer term, I'd prefer to sell by the core, since core-splitting
via hyperthreading doesn't make much sense.

The prefix "m5a" happens to be somewhat similar to the product code in
AWS we'd be targeting on Hetzner: an AMD Zen2 system with a 4GB * 1
core ratio.  But for demonstration, these are backed by whatever I
could buy off the auction today, which is Intel, though with the
correct (and relatively high) RAM ratio.  Alas, we're probably a few
days away from our AMD Zen2 EPYC servers being available.  And we'll
probably change our product coding as well; this is to help audiences
get the right general idea.