Avoid default routing in VM namespace
The default route here has bothered me for quite a while: while
developing I've introduced routing loops, and it makes it all too easy
to spill unencrypted intended-to-be network overlay traffic into the
host network namespace and then the ethernet card, if IPsec policies
are not set just right to skim off those packets and encrypt &
encapsulate them first.

As-is, as we never need to contact anything from a guest VM namespace
that's not an IPv6 GUA (Global Unicast Address), so go ahead and
itemize those explicitly.  Now, any otherwise leaked ULA address
traffic will find itself without a route, which is much better.

I gave this patch a try and it seems to have no ill effect on Internet
connectivity from the VM.