Revive explicit DNS server setting for dnsmasq
By default, this will ask `systemd-resolved` where to do the query,
which fails given that systemd configures `systemd-resolved` as yet
another level of recursive resolver, bound to localhost,
e.g. 127.0.0.53.  dnsmasq is in a separate namespace that does not
have the same localhost, so it cannot contact `systemd-resolved`, and
we probably don't want two levels of recursive resolution anyway.

So instead, install some datacenter-neutral-ish DNS servers.  These
are from Quad9.