Support a-priori net6
Using it looks like:

    Prog::Vm::HostNexus.assemble('169.150.215.90', net6: '2a02:6ea0:e80d::/64')

DataPacket has a different way of configuring IPv6 static routes that
we may see more of in the future.  The general way it works is the
interface has to respond to IPv6 NDP on a particular address, and the
result is the non-overlapping subnet is routed to that machine.

Because that non-overlapping subnet is not sound to compute from
anything automatically "learned" from the instance itself, it must be
programmed a priori.

And to do that, there need to be a few key relaxations in constraints
and a few branches to get around code assuming the style where the
network range is "learned" from the interface already configured.