Bring VmHost to 100% and fix collision resolution bugs
`==` does not do what a reasonable person would think they ought to do
with NetAddr IPv6 objects (it appears to be more like reference
equality).  So use `#cmp(o) == 0`.

Also, the re-run of obtaining a random subnet was never being
returned, so instead, the conflicting "proposal" would be returned
instead.  One way to fix this would be to use `return` (that was my
intention earlier) but I decided to use a trailing `case` instead for
no particular reason.

Finally, one invariant is removed, as there are cases where there's no
reserved subnet, for example, DataPacket statically routes an entire
/64 linked to a computer responding via NDP to a formally unrelated
/64 network.  As such, the host takes no bite out of the subnets
available to guests here.