Switch from Ractors to Threads for stall termination
Alas, Ractors cause some problems with non-deterministic coverage
results, that do not occur with Threads.  It looks suspiciously like
the Ractors are flushing coverage state to me.

I wrote a somewhat superficial bug report at
https://github.com/simplecov-ruby/simplecov/issues/1058 about this.

In the past, I have managed this problem with Threads, and they've
been fine, even if they widen the potential for a theoretical problem
of a thread not getting scheduled because of a bug in other code in
releasing the GVL.