Hoist strand updating from Prog::Base to Strand
As I was writing prog tests, I found that many tests could be written
without database interactions for some speedup, but some were made
unnecessarily brutal because base.rb made calls to
`Sequel::Model#update` directly.

So I hoist these up to the site where `FlowControl` exceptions are
rescued, and do the database modification at the last moment instead.
The resulting code is a bit less simple, so it's offered on the
principle of making the natural testing procedure a bit more
efficient.