Use SPDK to provide storage to VMs
We want to use SPDK for storage in the longer run. This allows to
have more control over storage, and provide features like snapshots,
encryption, and replication.

Before this PR, we just passed the $VmHome/boot.raw file as the
VM disk:

boot.raw <-> VMM

This PR changes that and adds SPDK as an intermediate:

boot.raw <-> SPDK <-(vhost protocol)-> VMM

Although this doesn't add a new functionality, but it paves the
way to modify the SPDK layer above to add other features in future.

Follow up items in next PRs:
* Record information about VMs storage topology in database
* Add encryption layer
* Add the possibility to provide raw NVMe disks directly to VMs