Enable autoload for clover app at dev
We have some `Unreloader.autoload` directives, but autoload isn't
enabled. It works similar to `require` when not enabled.
Autoload speeds things up at dev, or when only running a subset of
tests. If the constants are not accessed, you don’t need to pay the cost
of loading the related files.

Autoloading hash branches is little bit tricky. Of course Jeremy has
`autoload_hash_branches` plugin and it accepts directories too.
`autoload_hash_branch_dir` expects filename and hash branches
to be same. But filenames use `_` and hash branches use `-` as
separator. Because of that we autoload files one by one instead of using
`autoload_hash_branch_dir`.