Make storage size & id customizeable
Previously disk size was fixed at image_size+10G. This PR makes users able to choose the disk size. Users can do this by providing a `storage_size_gib` parameter to `Vm.assemble`:

```
> st = Prog::Vm::Nexus.assemble(mykey, nil, storage_size_gib: 30)
```

Also, we explicitly set disk id when creating the VM, so users can access disks using `/dev/disk/by-id/virtio-#{vm_name}_#{disk_index}`. For example, if a vm has a single disk, we will see the following from inside the VM. One id for the main disk, and 3 ids for each of its partitions.

```
$ ls /dev/disk/by-id/virtio-vm* -1
/dev/disk/by-id/virtio-vm3z0fs9_0
/dev/disk/by-id/virtio-vm3z0fs9_0-part1
/dev/disk/by-id/virtio-vm3z0fs9_0-part14
/dev/disk/by-id/virtio-vm3z0fs9_0-part15
```

Some notes on how the device_id is communicated between different components are as follows.

The device id in above is provided by the virtio device. Linux kernel extracts & uses it in virtblk_get_id function in virtio_blk driver [1].

For file backed disks which we used to use (not anymore), cloud-hypervisor sets this id as "$dev$rdev$inode" in [2].

For SPDK backed disks, SPDK sets device id to the actual bdev name in SPDK [3].

So in this patch we pass the device_id directly as SPDK device name. Because we use the same SPDK instance for all VMs on the same host, we require the device_id to be unique across all VMs in a host.

[1] https://github.com/torvalds/linux/blob/5f63595ebd82f56a2dd36ca013dd7f5ff2e2416a/drivers/block/virtio_blk.c#LL900C55-L900C61
[2] https://github.com/cloud-hypervisor/cloud-hypervisor/blob/95626ae56459f65aebd0029bb8e99dfd285575e5/block_util/src/lib.rs#L77
[3] https://github.com/spdk/spdk/blob/1334e3e40700b9f9f9c9d357bc748a33fb08797a/lib/vhost/vhost_blk.c#L625