Upgrade dnsmasq to version v2.89
Per usual, I prefer to use a commit SHA1 with a fsck to establish
something of a "trust on first use" relationship with the artifact.

I figured it'd be better to use a release rather than a commit I
selected in a mostly, but not entirely, arbitrary manner: I thought
there would be value for having new versions of dnsmasq for such an
important dependency, and a chance we'd want to make alterations or
pick up specific new changes.