Improve dnsmasq cloning for commits not at the tip
`git clone` with `depth 1` will only be able to handle commits that
co-incide with the default branch of the repository, which breaks the
moment the mirror is updated.

And unfortunately it doesn't look like `git clone` can accept a ref
directly for `--depth 1`, so instead use `git fetch` on a blank
repository, which can accept a ref and ask the remote computer for
only that reference and the objects it needs to check it out.