Get rid of cloud/cloud123 user
This was very useful back when we'd routinely break networking and
need to use serial teletype to figure out what was going on, but that
development mode is so much more rare now that it's acceptable to
require patching up the code in development to get access to the VM
this way.

It was adapted directly from the cloud-hypervisor test scripts.

Fixes #58