Remove runvm.rb
This may even work, and it's useful as it gets the VM to run outside
systemd in the foreground, but it's quite dangerous since there's no
interlock with a another running copy of the VM (e.g. via systemd, per
normal).

Given that the development path of runvm.rb is less intense than it
once was, let's get rid of it.  If reintroduced, it can be made safer.
It's also possible that running a systemd unit in a foregrounded,
teletype-attached mode may also work.