Convert implied hash interface to Struct
This implied interface is seen in a few places in application code,
and in many cases in the tests, but since it's a literal with no
interface checking, any one of these sites can fall out of
synchronization.  Because a lot of it is in tests, the result is bogus
tests.

Using a struct fixes this, although it winds up more wordy to
construct the literal in tests.

A positional argument style Struct would have been more terse, but I
find the potential for error in production code paths unsatisfying in
that case.