Expand NAT to include all private IPv4 ranges
The previous commit 6524645b650262d7efd1bdc5a3b411e7361ed356 ommits to itemize
private ipv4 Class A (10.0.0.0/8) and Class B (172.16.0.0/12) address ranges for
NAT. We are adding them to the list of NATted ranges because we want to support
them as part of the virtual networking setup. This means, via nftables, for an
outgoing packet, if the destination ip is not part of these subnets, we replace
the source ip addr with the public ip, becuase simply client wants to
communicate with public ip addresses. If not, we keep it as is and the
communication is handled by ipsec tunnels in later stages.