Add project switcher component to sidebar
We decided to show resources of a single project at a time on web
console similar to other providers.
First of all, it helps to not mixing production and development
resources. It feels safer while working on the list of resources.
It feels more intuitive in some cases.
We're able to add direct links to project's users, access policy,
billing pages to sidebar. Also add direct links to dashboard too. It
helps to navigate initial users easier.

Project switcher helps to change current project easily.