Introduce Page model
This commit introduces simple page model. At the moment, we are not concerning
about page's content. It only has a summary column as content. Our motivation
right now is providing enough baseline for deadline infrastructure. After that
it is possible to enrich the content of the pages to provide more information
to on-call engineers.