Pagerduty integration
This commit introduces Pagerduty integration. Added functionality only runs if
Config.pagerduty_key is set. Otherwise it would continue to create pages in the
database but does not/cannot trigger pages.

Added code for the Pagerduty integration is quite small and unintruisive, so
that it can be replaces with another provider if necessary. For the moment, our
expectation from on-call provider is quite minimal:
- Ability to trigger alerts (notification, e-mail, SMS, phone call)
- Ability to set on-call schedule and escalation policy

Both are part of core features of such providers, so it isn't big effort to
replace it.