Introduce label based deadline infrastructure
Deadlines are pretty key in our mode of operation to detect anomalies in the
system. Any operation that doesn't progress in a timely manner would create
a deadline and trigger an incident for on-call person to look at.

Initially we planned to put these deadlines only for the exit time of Progs.
That means if a Prog doesn't exit in a given time duration, it would create a
deadline. However that is bit restrictive, so instead we are introducing label
based deadlines. Label based deadlines would allow setting a deadline for Prog
to not just exit but also for reaching to an intermediate label.

Currently deadline related properties live in stack frames. It is possible
that number of deadline related fields increase over time. For example, it is
likely that we would want to trigger incident only for certain deadlines, which
would require adding deadline_severity field to the stack frames. I think it
would be better to create a deadline model at that time to not pollute frames
to much and then only keeping deadline_id in the frame.

As a side note, we have 3 strand.modified!(stack) calls in this PR. The reason
is that Sequel is only able to detect modifications if they are made using
column's setter direcly like "strand.stack = ...". It does not detect the
updates we made to stack elements because we are updating elements  of the
stack, not the stack itself. Thus, we are letting it know that strand column is
modified and the change needs to be persisted in the DB.