Track all files for actual code coverage
We are using Unreloader with autoload feature. It's not loading
files when file isn't used. So when a file not used in any test,
that file isn't shown in coverage report. But the file should be
in coverage report zero coverage.

`SimpleCov.track_files` helps to track all files.

This change decreased our current code coverage limits.