Move icons to separate file
There are several ways to show SVG in HTML. One of the recommended way is
showing SVG inline. We are doing it. But long SVG paths make hard to
maintain HTML files. Also some of the icons used multiple times.

I created an icon component and show icon depending on its name. So all
of the our SVG icons go to "icon.erb" file.

I excluded this file from erb_formatter because it couldn't format `when`
cases correctly.