Add missing nil check in register_deadline
Previously we didn't add nil check before parsing deadline_at, because we had
check on the deadline_target already. However for the deadlines we specify for
Prog exit, we put nil target. In such cases we need to nil check on deadline_at
as well. In the current logic, register_deadline only registers a deadline if
it satisfies any of the 3 conditions;
- There is no deadline set (determined by checking if deadline_at is nil)
- Current deadline is set for a further time than new deadline registration
- Current deadline is set for another target

Apart from that, we also start to clear pages if it exist and the target is
changed.