Use ${vendor}${year}-${pad}${core}c-${pad}${ram}r product notation
Previously, we used AWS-style names to convey a general idea to those
receiving a demonstration, but there are many problems with AWS-style
names that we don't need to repeat. In particular, they convey nothing
useful about the instance and sort badly, sending me off to the manual
with regularity for simple matters.

The results are like this:

    amd19-0.3c-1.3r
    amd19-0.3c-2.6r
    amd19-1c-4r
    amd19-1c-8r
    amd19-2c-8r
    amd19-2c-t16r
    amd19-4c-t16r
    amd19-4c-t32r
    amd19-8c-t32r
    amd19-8c-t64r
    amd19-t16c-t64r
    amd19-t16c-u128r
    amd19-t32c-u128r
    amd19-t32c-u256r
    amd19-t64c-u256r
    amd19-t64c-u512r
    amd19-u128c-u512r
    amd19-u128c-v1024r

A string like `amd19-2c-8r` means: an AMD chip, released in 2019, two
cores, eight gigabytes of ram.

A string like `amd19-u128c-u512r` means: an AMD chip, released in
2019, 128 cores, 512 gigabytes of ram.  Note the `u` character has no
meaning: it's only for sort assistance, explained in more detail
below.

A string like `amd19-0.3c-1.3r` is intended to be for a third of a
core, a product we don't support yet, but I wanted to include how it
might look.

The meaning of each part:

`c` is cores, `r` is gigabytes of ram.  We could omit these to be
positional like `amd19-2-4`, but I thought it'd invite errors as
people would swap the meaning of the two positions.

the `t` and `u`, `v` bytes seen with two, three, and four digit
numbers respectively are used to help the product strings sort.  They
have no other meaning.  No such sort assist byte is necessary for
decimalized numbers or single-digit numbers.

This is the strangest part of the design, introducing a meaningless
character to assist sorts.  I'm not convinced of it, but bad sorts on
product names have been a pebble in my shoe for a long time.  Normally
this is accomplished by zero-padding, but this would result in long
and hard to type product strings for small products, which tend to be
the most commonly used, e.g. a hypothetical `amd19-002c-0008r`.  I
thought this was intolerable.

== Comparisons to other vendors

In the form of shell expansion with a default: `${symbol:example}`

GCP is of the form:

    ${obscure-microarchitecture-string:t2d}-${ram-ratio-string:standard}-${cores}

Example: `t2d.standard-4`

Azure is of the form:

    Standard_${ram-ratio+manufacturer-char:D}#{vcpu:64}${other-char-modifiers:s}_${microarch_version:v5}

Example: `Standard_D64ds_v4`

I don't know why the `Standard_` shows up so often without variation
in Azure.  Another annoying quirk is how instance lines are referred
to like "Dds" but in a product name, the vcpu count interrupts, so you
can't copy and paste.

AWS is of the form:

    ${memory-ratio:m}${microarchish:5}${manufacturer:a}.${t-shirt-size}

Example: `m5a.xlarge`

== The Sort Assist Byte Range

The general idea here is to choose bytes that are:

* Contiguous (easy to write a regexp range against)
* Meaningless

So I wrote a table of the alphabet and assigned meaning to characters.
The range `[t-w]` stood out as a good range in the lack of meaning
present:

    a - arm, amd
    b
    c - core
    d - disk
    e
    f
    g - gigabyte
    h
    i - intel, instance
    j
    k
    l - thin when in non-monospace fonts
    m - memory, megabyte
    n - network
    o
    p
    q
    r - ram
    s - storage
    t
    u
    v
    w
    x - multiply?
    y
    z