Introduce billing rates
Rest of the usage record logic depends on product naming and requires some
modification on the DB side, which will be integrated with separate commit.

This commit introduces billing rates, which will be used as catalog for all
pricing calculations and invoices. Each place that needs to calculate a price
will be able to fetch unit price of particular resource from this table, and
then calculate the total price as unit_price * usage_amount * usage_duration.

For example, for VMs, if the a customer uses 4 cores from c5a family for 1000
minutes in hetzner-fsn1 region, total price would be 0.000171296 * 4 * 1000.
Unit of unit_price for VMs is minute, but each product can determine its own
unit. For disks, it could be GB or for a serverless product it could be number
of requests. It is possible to add human readable unit column so eliminate
potential confusion, but I'm leaving that out for now. We can add if we see
a need for that.

Initially, I started to build more general solution that would satisfy the
needs of many future products, but that proved itself quite challenging as
there are lots of unknown unknowns. Instead, I decided to build much more
simple and flexible model that can be extended as we learn more about the
product we are building.

Current model is quite flexible and allows numerous potential expansion paths.
For example, reserved instance pricing can be implemented easily by adding a
column.

Similarly price updates can be done quite easily by adding two new columns;
effective_since and effective_until. Then it is matter of ensuring invoice
generation and price calculator honours those dates. Alternative would be
generating new billing rates and depreciating the old ones, but updating
billing rate references everywhere requires manual database surgery and is
open to operator mistakes.