Archive deleted records to a separate table
This is prerequisite for audit records. We should be able to identify which VMs
and IP addresses used by which Projects/Accounts at a given time, so that we
can take necessary precautions against network abuse and spamming.

Since keeping historical records are useful for debugging as well, with this
commit we enable soft delete for not only VMs/IP addresses but for almost all
other models too. There are two exceptions at the moment;
- DeletedRecord: Not archived to prevent infinite loops, also archiving them
is not very meaningful.
- Semaphore: Not archived because we produce a lot of them.

Of course keeping historical records makes GDPR compliance a bit more difficult,
but that can be resolved by deleting very old records via batch processing after
certain duration.