Discourage using delete method
Using delete method while deleting Sequel models does not trigger destroy hooks,
which we use for archival. With an earlier commit, we already removed all delete
calls from our codebase. However, it is still possible that a delete call might
be introduced in the future or an operator might call delete manually by mistake.

To prevent such unintentinal delete calls, with this commit, we are prepending
Sequel::Dataset and Sequel::Model classes to overwrite the delete behaviour. Now
using delete method raises an error and suggests using destroy instead. If delete
is really necessary in the given context, it can be still called via force flag.

One ugly part of this commit is that in our version of delete, we check the call
stack to see if;
- delete is called from rodauth
- delete is called from destroy itself
to allow those uses of delete.