Remove Nic wait from Vm:Nexus
This commit remvoes the Nic waiting from Vm:Nexus. That's not necessary as we
don't essentially need nic to be in any specific state to create the VM. It's
enough that the Nic is created and private ip addresses are assigned. Waiting
for Nic which makes Nic to wait for PrivateSubnet can end-up in a race
condition for Subnets with >1 VM and another VM being provisioned.
Because, the moment Nic is created, it's a destination address for all the Nics
in the subnet, new Vm not being fully operable causes the tunnel creation
failure for all the other Nics. Since the Subnet is not in wait state, we also
start waiting for it to be functional before continuing with VM provisioning.
Therefore, we create a circular/never ending dependency. This commit breakes
that chain by simply continuing with VM provisioning. We are already triggering
a refresh_mesh after the VM provisioning, so we do not lose any functionality
doing so.