Upgrade SPDK to include the crypto fix
There used to be a bug in SPDK [1] which prevented us from using
encrypted volumes reliably. This was fixed by [2]. This PR changes
the SPDK package to a version which includes that fix.

For testing I tried the following in an encrypted VM without any
issues:

* The `dd` command in [1] which replicated the issue in the
  previous buggy version.
* apt update & upgrade & install postgres requirements.
* clone postgresql & build it & run it.
* create a table with 10m rows & run couple of aggregates.

I also tried the above steps in another unencrypted VM.

[1] https://github.com/spdk/spdk/issues/3076
[2] https://review.spdk.io/gerrit/c/spdk/spdk/+/19232