Switch product names to standard-(cores*2) notation
Why cores * 2? Because SMT features, as seen on Intel and AMD products
for some year now make this equal to the number of "CPUs" or "vCPUs"
when exposed to the guest operating system.

The theory is that an integer in the product code will be compared to
vCPU references similar products on various data tables and so forth
on The Internet, or more to the point, shouldn't be a smaller integer
than those references for the same hardware.  Note, I do not subscribe
to theory's assigned magnitude of this, and have my objections to what
it does to the cogency of product comparison as written below, but
nonetheless, I've written this operating under this theory.

The ugliness arrives when we consider SMT-off configurations or
relatively recent addition of ARM processors that do not implement
SMT.  If the product scale factor was taken to describe vCPUs, we'd
end up the situation that GCP and Azure are in: t2d.standard-1 and
t2a.standard-1 are one vCPU in SMT-off configuration, and thus a
single core, and most directly comperable to n2d.standard-2.  The
customer is left the job of being SMT-aware and scaling the model
names.

Azure did something rather strange to deal with this, where a
"Standard_E2" is 2 vCPUs on one core via SMT and 16 gigabytes of
memory, but "Standard_E2p" on ARM is 2 cores and 16 gigabytes of
memory, a more compute-rich ratio by quite a bit.

One solution, absent indexing the product by cores directly, which
produces an integer that could be misconstrued and deemed too small,
is to double the core number in all circumstances.  The idea of a
popular architecture with three or more threads per core seems
unlikely at this point, and in any case it'd be an emergent
architecture where we could be forgiven for an awkward numbering
system.  And it solves the problem for allowing someone to scan
laterally from "standard-2" and a hypothetical "standard-2-arm".  Even
if standard-2-arm has a single core.

Of course, there's nothing stopping us from deciding to really make
the product scale factor relate to vCPUs, that is, to have integers of
larger magnitudes for architectures with SMT, and maybe if people do
prove to be uniformly alert about SMT and find `standard-2-arm`
misleading, we can elect to do things that way.

If it were up to me, I'd index the model numbers on cores, and in
addition to cleaning all that up, would also attain the aesthetically
pleasing "standard-1".