Reap should continue deleting, not destroy
The commit 06b9eb8 changed the reap behavior to destroy but .destroy only returns
the row count when used with returning. On the other hand, .delete returns the
objects deleted in hash form which we use later in the host_nexus. So, the
VmHost setup was broken. This commit fixes that.